switch from using int to long in IDataProvider and the cascade from that Issue: WICKET-1175
git-svn-id: https://svn.apache.org/repos/asf/wicket/trunk@1203322 13f79535-47bb-0310-9956-ffa450edef68

+++ testing/wicket-threadtest/src/main/java/org/apache/wicket/threadtest/apps/app1/ContactsDatabase.java
@@ -83,9 +83,9 @@ public class ContactsDatabase
 	 * @param sortAsc
 	 * @return list of contacts
 	 */
-	public List<Contact> find(int first, int count, String sortProperty, boolean sortAsc)
+	public List<Contact> find(long first, long count, String sortProperty, boolean sortAsc)
 	{
-		return getIndex(sortProperty, sortAsc).subList(first, first + count);
+		return getIndex(sortProperty, sortAsc).subList((int)first, (int)(first + count));
 	}
 
 	/**
