+++ wicket-core/src/main/java/org/apache/wicket/markup/repeater/data/DataViewBase.java
@@ -69,7 +69,7 @@ public abstract class DataViewBase<T> extends AbstractPageableView<T>
 
 
 	@Override
-	protected final Iterator<IModel<T>> getItemModels(int offset, int count)
+	protected final Iterator<IModel<T>> getItemModels(long offset, long count)
 	{
 		return new ModelIterator<T>(internalGetDataProvider(), offset, count);
 	}
@@ -86,8 +86,8 @@ public abstract class DataViewBase<T> extends AbstractPageableView<T>
 	{
 		private final Iterator<? extends T> items;
 		private final IDataProvider<T> dataProvider;
-		private final int max;
-		private int index;
+		private final long max;
+		private long index;
 
 		/**
 		 * Constructor
@@ -99,7 +99,7 @@ public abstract class DataViewBase<T> extends AbstractPageableView<T>
 		 * @param count
 		 *            max number of items to return
 		 */
-		public ModelIterator(IDataProvider<T> dataProvider, int offset, int count)
+		public ModelIterator(IDataProvider<T> dataProvider, long offset, long count)
 		{
 			this.dataProvider = dataProvider;
 			max = count;
@@ -137,7 +137,7 @@ public abstract class DataViewBase<T> extends AbstractPageableView<T>
 	}
 
 	@Override
-	protected final int internalGetItemCount()
+	protected final long internalGetItemCount()
 	{
 		return internalGetDataProvider().size();
 	}
