+++ wicket-core/src/test/java/org/apache/wicket/markup/html/navigation/paging/AbstractPagingNavigationTest.java
@@ -34,13 +34,13 @@ public abstract class AbstractPagingNavigationTest extends WicketTestCase
 		private static final long serialVersionUID = 1L;
 
 		/** expected page which is set by the link. */
-		private int expectedPage = 0;
+		private long expectedPage = 0;
 
 		/**
 		 * @see IPageable#getCurrentPage()
 		 */
 		@Override
-		public int getCurrentPage()
+		public long getCurrentPage()
 		{
 			return currentpage;
 		}
@@ -49,7 +49,7 @@ public abstract class AbstractPagingNavigationTest extends WicketTestCase
 		 * @see IPageable#setCurrentPage(int)
 		 */
 		@Override
-		public void setCurrentPage(int page)
+		public void setCurrentPage(long page)
 		{
 			assertEquals("setCurrentPage", expectedPage, page);
 		}
@@ -58,7 +58,7 @@ public abstract class AbstractPagingNavigationTest extends WicketTestCase
 		 * @see org.apache.wicket.markup.html.navigation.paging.IPageable#getPageCount()
 		 */
 		@Override
-		public int getPageCount()
+		public long getPageCount()
 		{
 			return pagecount;
 		}
@@ -75,8 +75,8 @@ public abstract class AbstractPagingNavigationTest extends WicketTestCase
 	}
 
 	/** mock page count. */
-	protected int pagecount = 0;
+	protected long pagecount = 0;
 
 	/** mock current page. */
-	protected int currentpage = 0;
+	protected long currentpage = 0;
 }
